Refactor of Purrlaroid Component
