+++ components/purrlaroid/index.ts
@@ -1,1 +1,7 @@
-export * from "./Purrlaroid";
+import PurrlaroidRoot from "./PurrlaroidRoot";
+import PurrlaroidTitle from "./PurrlaroidTitle";
+
+export const Purrlaroid = {
+  Root: PurrlaroidRoot,
+  Title: PurrlaroidTitle,
+};
